requirements.txt
anyio==3.3.4
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
fastapi==0.70.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
idna==3.3
importlib-metadata==4.8.2
mariadb==1.0.8
passlib==1.7.4
psycopg2==2.9.2
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
PyMySQL==1.0.2
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.16.0
typing-extensions==4.0.0
uvicorn==0.15.0
zipp==3.6.0
